Project Description
Working under Professor Bhat of the University of Illinois Urbana-Champaign, my group and I were able to prepare a cohesive research report that analyzed the different factors on both the student and family level that negatively/positively impact student academic performance in the European and Middle Eastern Countries(Kaggle Datasets). Throughout this report, we showcased how we established our problem of focus, different preliminary visualizations to analyze the data, and a Machine Learning-based approach to distinguishing which variables have the most impact on student academics.